Correcting compiler warnings in Framework core code
Mainly about unused variables in the code, initialisation sequence of class
members and comparison between signed and unsigned. Some warnings remain as these
would require additional code changes.

Also correcting a minor bug in a unit test which was indicated by compiler
warning 'unused variable'.

+++ Framework/Core/include/Framework/TableBuilder.h
@@ -241,7 +241,7 @@ class TableBuilder
   }
 
   template <typename... ARGS>
-  auto makeBuilders(std::vector<std::string> const& columnNames, size_t nRows)
+  auto makeBuilders(std::vector<std::string> const& columnNames, int nRows)
   {
     using BuildersTuple = typename std::tuple<std::unique_ptr<typename BuilderTraits<ARGS>::BuilderType>...>;
     mSchema = std::make_shared<arrow::Schema>(TableBuilderHelpers::makeFields<ARGS...>(columnNames));
