+++ Framework/Core/test/benchmark_TableBuilder.cxx
@@ -34,7 +34,7 @@ static void BM_TableBuilderScalar(benchmark::State& state)
   for (auto _ : state) {
     TableBuilder builder;
     auto rowWriter = builder.persist<float>({ "x" });
-    for (size_t i = 0; i < state.range(0); ++i) {
+    for (int i = 0; i < state.range(0); ++i) {
       rowWriter(0, 0.f);
     }
     auto table = builder.finalize();
@@ -50,7 +50,7 @@ static void BM_TableBuilderScalarPresized(benchmark::State& state)
   for (auto _ : state) {
     TableBuilder builder;
     auto rowWriter = builder.preallocatedPersist<float>({ "x" }, state.range(0));
-    for (size_t i = 0; i < state.range(0); ++i) {
+    for (int i = 0; i < state.range(0); ++i) {
       rowWriter(0, 0.f);
     }
     auto table = builder.finalize();
@@ -68,7 +68,7 @@ static void BM_TableBuilderScalarBulk(benchmark::State& state)
   for (auto _ : state) {
     TableBuilder builder;
     auto bulkWriter = builder.bulkPersist<float>({ "x" }, state.range(0));
-    for (size_t i = 0; i < state.range(0) / chunkSize; ++i) {
+    for (int i = 0; i < state.range(0) / chunkSize; ++i) {
       bulkWriter(0, chunkSize, buffer.data());
     }
     auto table = builder.finalize();
@@ -83,7 +83,7 @@ static void BM_TableBuilderSimple(benchmark::State& state)
   for (auto _ : state) {
     TableBuilder builder;
     auto rowWriter = builder.persist<float, float, float>({ "x", "y", "z" });
-    for (size_t i = 0; i < state.range(0); ++i) {
+    for (int i = 0; i < state.range(0); ++i) {
       rowWriter(0, 0.f, 0.f, 0.f);
     }
     auto table = builder.finalize();
@@ -98,7 +98,7 @@ static void BM_TableBuilderSimple2(benchmark::State& state)
   for (auto _ : state) {
     TableBuilder builder;
     auto rowWriter = builder.persist<float, float, float>({ "x", "y", "z" });
-    for (size_t i = 0; i < state.range(0); ++i) {
+    for (int i = 0; i < state.range(0); ++i) {
       rowWriter(0, 0.f, 0.f, 0.f);
     }
     auto table = builder.finalize();
@@ -122,7 +122,7 @@ static void BM_TableBuilderSoA(benchmark::State& state)
   for (auto _ : state) {
     TableBuilder builder;
     auto rowWriter = builder.cursor<TestVectors>();
-    for (size_t i = 0; i < state.range(0); ++i) {
+    for (int i = 0; i < state.range(0); ++i) {
       rowWriter(0, 0.f, 0.f, 0.f);
     }
     auto table = builder.finalize();
@@ -137,7 +137,7 @@ static void BM_TableBuilderComplex(benchmark::State& state)
   for (auto _ : state) {
     TableBuilder builder;
     auto rowWriter = builder.persist<int, float, std::string, bool>({ "x", "y", "s", "b" });
-    for (size_t i = 0; i < state.range(0); ++i) {
+    for (int i = 0; i < state.range(0); ++i) {
       rowWriter(0, 0, 0., "foo", true);
     }
     auto table = builder.finalize();
